requirements.txt
langchain
openai
langchain-openai
pydantic==2.6.1
faiss-cpu==1.7.4
streamlit==1.29.0
yfinance==0.2.31
python-dotenv==1.0.0
